server/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

langchain==0.2.14
langchain-google-genai==1.0.10
